feat[staging-provider]: Remove workspace provider configuration, replace by specific methods
