--- libs/staging-provider/src/Workspace/Credentials/NoCredentialsProvider.php
@@ -1,16 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Keboola\StagingProvider\Workspace\Credentials;
-
-use Keboola\StagingProvider\Exception\StagingProviderException;
-use Keboola\StagingProvider\Workspace\WorkspaceInterface;
-
-class NoCredentialsProvider implements WorkspaceCredentialsProviderInterface
-{
-    public function provideCredentials(WorkspaceInterface $workspace): array
-    {
-        throw new StagingProviderException('Credentials are not available');
-    }
-}
