--- libs/staging-provider/src/Workspace/ProviderConfig/WorkspaceConfigInterface.php
@@ -1,9 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Keboola\StagingProvider\Workspace\ProviderConfig;
-
-interface WorkspaceConfigInterface
-{
-}
